Rename js or jsx to tsx
- Migrate project to typescript.
