+++ todolist-frontend/src/App.tsx
@@ -1,9 +1,8 @@
-import React from 'react';
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
 
-import Auth from './components/Auth/Auth';
+import Auth from './pages/Auth/Auth';
 import './App.scss';
-import Dashboard from './components/Dashboard/Dashboard';
+import Dashboard from './pages/Dashboard/Dashboard';
 
 const App = () => {
